Added base class for cubesphere output
